utils-gradient-compositon.lisp: allow use of gradients instead of colours as fill and stroke targets for wand drawing functions
* utils-gradient-compositon: (call-draw-function-with-gradient-composition,
with-gradient-composition): wrappers around wand drawing functions
which allow the user to optionally specify a fill-gradient or a stroke-gradient
(either a MagickWand object or a string).

1. blacken out the given gradient image 2. use the draw function to
prepare an appropriate b&w mask overlay - black is transparent and
white is what is drawn by the drawing function. 3. composite the
overlay over the gradient with "CopyOpacity", and composite the
gradient over the background with "Over".

with some help from the friendly IM folks at
https://github.com/ImageMagick/ImageMagick/discussions/8449#discussioncomment-15014963